requirements.txt
python-telegram-bot==20.4
yt-dlp==2024.2.11
requests==2.31.0
